Use signed operator fields for SCP search state
SCHTAB logic and boolop store small search operator codes, not sizes.
The boolop field also uses -1 as the transient "no relational operator"
sentinel while parsing IF and ASSERT search expressions.

Use int for these fields so the type matches the value domain and the
sentinel checks do not trigger sign-comparison warnings.

+++ src/core/sim_defs.h
@@ -739,8 +739,8 @@ struct MTAB {
 /* Search table */
 
 struct SCHTAB {
-    size_t              logic;                          /* logical operator */
-    size_t              boolop;                         /* boolean operator */
+    int                 logic;                          /* logical operator */
+    int                 boolop;                         /* boolean operator */
     uint32              count;                          /* value count in mask and comp arrays */
     t_value             *mask;                          /* mask for logical */
     t_value             *comp;                          /* comparison for boolean */
